fix: added first part with warnings
