+++ uui-components/src/widgets/AvatarStack.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import cx from 'classnames';
 import { IHasCX, IHasRawProps } from '@epam/uui-core';
 import { Avatar } from './Avatar';
-import { FlexRow } from '../';
+import { FlexRow } from '../layout';
 import css from './AvatarStack.module.scss';
 
 export interface AvatarStackProps extends IHasCX, IHasRawProps<React.HTMLAttributes<HTMLDivElement>> {
