+++ uui-core/src/services/ModalContext.ts
@@ -2,8 +2,7 @@ import clone from 'lodash.clone';
 import { BaseContext } from './BaseContext';
 import { LayoutContext } from './LayoutContext';
 import * as React from 'react';
-import { IModal } from '../types/props';
-import { IModalContext } from '../types/contexts';
+import { IModal, IModalContext } from '../types';
 
 export interface ModalComponentProps<TParameters, TResult> {
     parameters?: TParameters;
