+++ uui-core/src/services/dnd/DndActor.tsx
@@ -1,7 +1,10 @@
 import * as React from 'react';
 import {
-    IDndActor, UuiContexts, DropPosition, AcceptDropParams, DndActorRenderParams, DropPositionOptions, DndContextState,
-} from '../../types';
+    IDndActor, DropPosition, AcceptDropParams, DndActorRenderParams, DropPositionOptions,
+} from '../../types/dnd';
+
+import { UuiContexts, DndContextState } from '../../types/contexts';
+
 import {
     isEventTargetInsideDraggable,
     isEventTargetInsideInput,
