+++ uui-core/src/types/contexts.ts
@@ -1,16 +1,14 @@
 import { Link, LayoutLayer } from './objects';
 import * as PropTypes from 'prop-types';
 import { IModal, INotification } from './props';
-import {
-    FileUploadOptions,
-    FileUploadResponse,
-    SkinContext,
-    ModalOperation,
-    NotificationOperation,
-    IHistory4,
-    Lock,
-    TMouseCoords,
-} from '../services';
+import { TMouseCoords } from '../services/dnd/DndContext';
+import { Lock } from '../services/LockContext';
+import { IHistory4 } from '../services/routing/HistoryAdaptedRouter';
+import { NotificationOperation } from '../services/NotificationContext';
+import { SkinContext } from '../services/SkinContext';
+import { ModalOperation } from '../services/ModalContext';
+
+import { FileUploadOptions } from '../services/ApiContext';
 
 export interface IBaseContext<TState = {}> {
     subscribe(handler: (state: TState) => void): void;
