+++ uui-editor/src/plugins/imagePlugin/ImageBlock.tsx
@@ -12,7 +12,7 @@ import invert from 'lodash.invert';
 
 import css from './ImageBlock.module.scss';
 import { ImgToolbar } from './Toolbar';
-import { IImageElement, PlateImgAlign, SlateImgAlign } from './imagePlugin';
+import { IImageElement, PlateImgAlign, SlateImgAlign } from '../../types';
 
 interface UpdatingProps { width?: number | string, align?: SlateImgAlign }
