+++ uui-editor/src/plugins/imagePlugin/Toolbar.tsx
@@ -9,7 +9,7 @@ import { ReactComponent as AlignRight } from '../../icons/align-right.svg';
 import { ToolbarButton } from '../../implementation/ToolbarButton';
 
 import css from './ImageBlock.module.scss';
-import { PlateImgAlign } from './imagePlugin';
+import { PlateImgAlign } from '../../types';
 
 export function ImgToolbar({
     align,
