+++ uui-editor/src/plugins/imagePlugin/imagePlugin.tsx
@@ -8,36 +8,13 @@ import { Image } from './ImageBlock';
 
 import { ToolbarButton } from '../../implementation/ToolbarButton';
 
-import { PlateEditor, TElement, createPluginFactory, focusEditor, getBlockAbove, insertEmptyElement, insertNodes } from '@udecode/plate-common';
+import { PlateEditor, createPluginFactory, focusEditor, getBlockAbove, insertEmptyElement, insertNodes } from '@udecode/plate-common';
 import { TImageElement, captionGlobalStore } from '@udecode/plate-media';
 import isHotkey from 'is-hotkey';
 import { ReactComponent as ImageIcon } from '../../icons/image.svg';
 import { PARAGRAPH_TYPE } from '../paragraphPlugin/paragraphPlugin';
-import { FileUploadResponse } from '@epam/uui-core';
 
-export type PlateImgAlign = 'left' | 'center' | 'right';
-export type SlateImgAlign = 'align-left' | 'align-right' | 'align-center';
-export type SlateImageSize = { width: number, height: number | string };
-
-type SlateImageData = {
-    imageSize: SlateImageSize;
-    align: SlateImgAlign;
-} & Partial<(File | FileUploadResponse)>;
-
-export interface SlateProps {
-    data: SlateImageData;
-}
-
-export interface PlateProps {
-    url: string;
-    align?: PlateImgAlign;
-    width?: number;
-}
-
-export interface IImageElement extends TElement, PlateProps, SlateProps {}
-
-export const IMAGE_PLUGIN_KEY = 'image';
-export const IMAGE_PLUGIN_TYPE = 'image';
+import { IMAGE_PLUGIN_TYPE, IMAGE_PLUGIN_KEY } from '../../types';
 
 export const imagePlugin = () => {
     const createImagePlugin = createPluginFactory({
