+++ uui-editor/src/plugins/uploadFilePlugin/file_uploader.ts
@@ -1,6 +1,5 @@
-import { IMAGE_PLUGIN_TYPE } from '../imagePlugin/imagePlugin';
 import { ATTACHMENT_PLUGIN_TYPE } from '../attachmentPlugin/attachmentPlugin';
-import { IFRAME_PLUGIN_TYPE } from '../iframePlugin/iframePlugin';
+import { IFRAME_PLUGIN_TYPE, IMAGE_PLUGIN_TYPE } from '../../types';
 import { useCallback } from 'react';
 import type { FileUploadResponse } from '@epam/uui-core';
 import { PlateEditor, getPlugin, KEY_INSERT_DATA, deleteBackward, insertEmptyElement } from '@udecode/plate-common';
