+++ uui/components/forms/useForm.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import {
     useUuiContext, UseFormProps, useForm as uuiUseForm, INotification,
 } from '@epam/uui-core';
-import { Text, RichTextView, WarningNotification } from '..';
+import { WarningNotification } from '../overlays';
+import { Text, RichTextView } from '../typography';
 import { ConfirmationModal } from '../overlays/ConfirmationModal';
 import { i18n } from '../../i18n';
